get rid of code copy of differential_ik
